Create the CMake file, re-organized project and new features
I created a CMakeLists that builds both the lib and unit tests.

You can now parse multiple flag as one like -abc = -a -b -c

Add a few unit tests, but there are still many missing.

+++ Js_CLI_Parser/include/Option.h
@@ -8,7 +8,7 @@ public:
 
 	const std::string& GetShortName() const;
 	const std::string& GetLongName() const;
-	int GetArity() const;
+	size_t GetArity() const;
 
 private:
 	std::string m_shortName;
